refactor(test): audit divergent_child.rs — dissolve, move to proof.rs
- All 3 tests exercise prove() and Proof::value_digest(), not change proofs
- Move to merkle/tests/proof.rs where other proof tests live
- Strip duplicate change proof round-trip from test 1 (covered by
  edge_cases test_boundary_child_gap_closed_for_start_key)
- Add crafted prefix to test_incomplete_exclusion_proof_rejected
- Delete divergent_child.rs and remove mod divergent_child from mod.rs
