+++ firewood/src/merkle/tests/change/mod.rs
@@ -6,7 +6,6 @@ use crate::api::{self, BatchOp, Db as DbTrait, DbView, FrozenChangeProof, Propos
 use crate::db::{Db, DbConfig};
 use crate::merkle::verify_change_proof_root_hash;
 use crate::{ChangeProofVerificationContext, verify_change_proof_structure};
-use firewood_storage::PathComponentSliceExt;
 
 // ── Test infrastructure ────────────────────────────────────────────────────
 
@@ -79,7 +78,6 @@ pub(super) fn verify_and_check(
 mod adversarial;
 mod attack;
 mod bounds;
-mod divergent_child;
 mod edge_cases;
 // Empty start trie tests require ethhash: without it, the empty trie has no
 // root hash, so change proofs from an empty database can't be generated.
